+++ metapackages/auto-instrumentations-node/CHANGELOG.md
@@ -46,6 +46,31 @@
     * @opentelemetry/instrumentation-pino bumped from ^0.31.0 to ^0.32.0
     * @opentelemetry/instrumentation-redis-4 bumped from ^0.32.0 to ^0.33.0
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-express bumped from ^0.31.1 to ^0.31.2
+
+## [0.33.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-node-v0.32.1...auto-instrumentations-node-v0.33.0) (2022-09-27)
+
+
+### Features
+
+* add dataloader instrumentation ([#1171](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1171)) ([3898b11](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/3898b11800f857c75c286f22c4633b5baf4e1f84))
+* add mongoose instrumentation ([#1131](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1131)) ([b35277b](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/b35277bb5fc66910e8942bc0b64347b68ecffa26))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-aws-sdk bumped from ^0.9.1 to ^0.9.2
+    * @opentelemetry/instrumentation-dataloader bumped from ^0.1.0 to ^0.2.0
+    * @opentelemetry/instrumentation-fastify bumped from ^0.29.0 to ^0.30.0
+    * @opentelemetry/instrumentation-mongodb bumped from ^0.32.0 to ^0.32.1
+    * @opentelemetry/instrumentation-mongoose bumped from ^0.30.0 to ^0.31.0
+
 ## [0.32.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-node-v0.31.2...auto-instrumentations-node-v0.32.0) (2022-09-02)